requirements.txt
google-adk[bigquery]>=1.14
python-dotenv>=1.0.1
pandas>=2.3.0
db-dtypes>=1.4.2
google-cloud-bigquery>=3.0.0
